requirements.txt
# 核心依赖
langchain>=0.1.0,<0.2.0
langchain-openai>=0.0.5
langchain-community>=0.0.13
openai>=1.10.0,<2.0.0
# 阿里云支持
dashscope>=1.17.0
langchain-community[llms]>=0.0.13

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.5.0,<3.0.0

# Web框架
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
streamlit>=1.28.0

# 数据库
sqlalchemy>=2.0.0,<3.0.0

# 其他工具
python-dotenv>=1.0.0
requests>=2.25.0
plotly>=5.0.0
pytest>=7.0.0
httpx>=0.24.0